+++ .github/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Maintenance Changelog
 This file contains update logs for this project. The top may contain a `Unreleased` section, gathering update logs for a future update during development.
 
+## 3.0.2 - May 17th 2019
+### Changed
+* Added config option `timerspecific-pingmessages` to have specific pingmessages shown when endtimers are run
+  * You can also set them ingame by using `/maintenance setmotd timer <index> ...` instead of `/maintenance setmotd <index> ...`, along with `/maintenance removemotd timer <index>` and `/maintenance motd timer`
+* Made dumps a little prettier
+* Support Sponge 8.0.0-SNAPSHOT
+* Support latest text/Velocity (now only supports Velocity builds from May 7th upwards)
+---
 ## 3.0.1 - April 17th 2019
 ### Changed
 * Created an addon to have PlaceholderAPI placeholders on Spigot when running the Maintenance plugin on Bungee (see [**here**](https://github.com/KennyTV/Maintenance/wiki/MaintenanceAddon))